requirements.txt
aws-cdk-lib==2.54.0
black==22.12.0
boto3==1.26.27
constructs>=10.0.0,<11.0.0
